Fixed a lots of bugs inside the cg.jl code
There was a lots of bugs, many of them due to typos, that now
are fixed. I have checked the whole code and, actually,
there was a bug in the precondictioner application and in the
Ap matrix building routine.

I have test the ConjGrad package with a strip2-like tod
and it's work fine.